Extract internal/configgen with exported Generate API
The pipeline-rendering and atomic-write logic moves out of
internal/cmd/cmdconfig and into a new internal/configgen package that
exports:

    configgen.Generate(ctx, dir, result *reposcan.Result) error

This is the seam an upcoming onboarding wrapper will use: it can run
its own scan and tests in-process, then call configgen.Generate to
write the config without re-scanning. The function is idempotent —
if .circleci/config.yml already exists, it prints "Using existing
config" and returns nil, matching the WEBXP-990 contract.

internal/cmd/cmdconfig keeps the CLI-specific work: positional path
validation, the pre-scan skip-check (so a no-op re-run never hits
the network), invoking the scanner, wrapping scanner errors, and
delegating to configgen.Generate. The YAML structs, renderConfig,
writeConfigAtomic, and the rename test seam all moved with the
logic.

Tests for YAML output, fallback template, and temp-file cleanup
move to internal/configgen. cmdconfig keeps tests for the path
validation, skip integration, scan invocation, and scanner-error
wrapping.
