+++ internal/cmd/cmdconfig/generate.go
@@ -23,7 +23,6 @@
 package cmdconfig
 
 import (
-	"bytes"
 	"context"
 	"fmt"
 	"os"
@@ -31,25 +30,18 @@ import (
 
 	"github.com/MakeNowJust/heredoc"
 	"github.com/spf13/cobra"
-	"gopkg.in/yaml.v3"
 
+	"github.com/CircleCI-Public/circleci-cli/internal/configgen"
 	clierrors "github.com/CircleCI-Public/circleci-cli/internal/errors"
 	"github.com/CircleCI-Public/circleci-cli/internal/iostream"
 	"github.com/CircleCI-Public/circleci-cli/internal/reposcan"
 )
-
-const generatedHeader = "# Generated by circleci config generate\n"
 
 // scan is the package-level seam for repository scanning. Tests swap it via
 // SetScanForTest in export_test.go.
 var scan = func(ctx context.Context, dir string) (*reposcan.Result, error) {
 	return reposcan.NewDefaultScanner().Scan(ctx, dir)
 }
-
-// rename is the package-level seam for finalizing the atomic write. Tests
-// swap it via SetRenameForTest to inject failures and confirm the temp-file
-// cleanup path works.
-var rename = os.Rename
 
 func newGenerateCmd() *cobra.Command {
 	cmd := &cobra.Command{
@@ -104,6 +96,9 @@ func runGenerate(cmd *cobra.Command, args []string) error {
 		).WithExitCode(clierrors.ExitBadArguments)
 	}
 
+	// Skip-check before scan so a no-op re-run never makes a network call.
+	// configgen.Generate also performs this check for callers that bypass
+	// the CLI flow.
 	configPath := filepath.Join(dir, ".circleci", "config.yml")
 	if _, err := os.Stat(configPath); err == nil {
 		iostream.ErrPrintf(ctx, "%s Using existing config at %s\n",
@@ -123,158 +118,5 @@ func runGenerate(cmd *cobra.Command, args []string) error {
 		).WithExitCode(clierrors.ExitGeneralError)
 	}
 
-	reposcan.Render(ctx, result)
-
-	body, err := renderConfig(result)
-	if err != nil {
-		return clierrors.New(
-			"config_generate.render_failed",
-			"Could not render config",
-			fmt.Sprintf("YAML marshaling failed: %s.", err),
-		).WithExitCode(clierrors.ExitGeneralError)
-	}
-
-	if err := writeConfigAtomic(configPath, body); err != nil {
-		return clierrors.New(
-			"config_generate.write_failed",
-			"Could not write config",
-			fmt.Sprintf("Failed to write %s: %s.", configPath, err),
-		).WithExitCode(clierrors.ExitGeneralError)
-	}
-
-	iostream.ErrPrintf(ctx, "%s Generated %s\n", iostream.SymbolOK(ctx), configPath)
-	return nil
-}
-
-// renderConfig builds the pipeline YAML body for a scan result. When the scan
-// produced no usable detection (IsEmpty), a generic cimg/base:stable template
-// is emitted with a placeholder build step. Output is deterministic: struct
-// field order controls top-level layout and yaml.v3 sorts map keys
-// alphabetically.
-func renderConfig(r *reposcan.Result) ([]byte, error) {
-	var image string
-	steps := []step{{Checkout: true}}
-
-	if r.IsEmpty() {
-		image = "cimg/base:stable"
-		steps = append(steps, step{Run: &runStep{
-			Name:    "build",
-			Command: `echo "Add your build steps here"`,
-		}})
-	} else {
-		image = r.Image
-		if r.ImageVersion != "" {
-			image = r.Image + ":" + r.ImageVersion
-		}
-		for _, s := range r.Setup {
-			steps = append(steps, step{Run: &runStep{Name: s.Name, Command: s.Command}})
-		}
-	}
-
-	p := pipeline{
-		Version: "2.1",
-		Jobs: map[string]pipelineJob{
-			"build": {
-				Docker: []dockerImage{{Image: image}},
-				Steps:  steps,
-			},
-		},
-		Workflows: map[string]workflow{
-			"build": {Jobs: []string{"build"}},
-		},
-	}
-
-	var buf bytes.Buffer
-	buf.WriteString(generatedHeader)
-
-	enc := yaml.NewEncoder(&buf)
-	enc.SetIndent(2)
-	if err := enc.Encode(p); err != nil {
-		return nil, err
-	}
-	if err := enc.Close(); err != nil {
-		return nil, err
-	}
-	return buf.Bytes(), nil
-}
-
-// writeConfigAtomic writes body to path via a temp file in the same directory
-// followed by rename. The .circleci directory is created if missing. If any
-// step after the temp file is created fails, the temp file is removed so the
-// .circleci/ directory is never left littered with config.*.yml.tmp files.
-func writeConfigAtomic(path string, body []byte) error {
-	if err := os.MkdirAll(filepath.Dir(path), 0o755); err != nil { //#nosec:G301 // .circleci/ is a repo-shared directory, world-readable like the surrounding workspace
-		return err
-	}
-
-	tmp, err := os.CreateTemp(filepath.Dir(path), "config.*.yml.tmp")
-	if err != nil {
-		return err
-	}
-
-	committed := false
-	defer func() {
-		if !committed {
-			_ = os.Remove(tmp.Name())
-		}
-	}()
-
-	if _, err := tmp.Write(body); err != nil {
-		_ = tmp.Close()
-		return err
-	}
-	if err := tmp.Sync(); err != nil {
-		_ = tmp.Close()
-		return err
-	}
-	if err := tmp.Close(); err != nil {
-		return err
-	}
-	if err := os.Chmod(tmp.Name(), 0o644); err != nil { //#nosec:G302 // config.yml is intended to be committed alongside the rest of .circleci/
-		return err
-	}
-	if err := rename(tmp.Name(), path); err != nil {
-		return err
-	}
-	committed = true
-	return nil
-}
-
-type pipeline struct {
-	Version   string                 `yaml:"version"`
-	Jobs      map[string]pipelineJob `yaml:"jobs"`
-	Workflows map[string]workflow    `yaml:"workflows"`
-}
-
-type pipelineJob struct {
-	Docker []dockerImage `yaml:"docker"`
-	Steps  []step        `yaml:"steps"`
-}
-
-type dockerImage struct {
-	Image string `yaml:"image"`
-}
-
-type workflow struct {
-	Jobs []string `yaml:"jobs"`
-}
-
-// step is either a bare "checkout" string or a "run" map. The custom
-// MarshalYAML selects the right form so the rendered YAML matches the
-// canonical CircleCI shape.
-type step struct {
-	Checkout bool
-	Run      *runStep
-}
-
-type runStep struct {
-	Name    string `yaml:"name"`
-	Command string `yaml:"command"`
-}
-
-func (s step) MarshalYAML() (interface{}, error) {
-	if s.Checkout {
-		return "checkout", nil
-	}
-	return map[string]*runStep{"run": s.Run}, nil
+	return configgen.Generate(ctx, dir, result)
 }
