+++ internal/configgen/configgen.go
@@ -1,0 +1,223 @@
+// Copyright (c) 2026 Circle Internet Services, Inc.
+//
+// Permission is hereby granted, free of charge, to any person obtaining a copy
+// of this software and associated documentation files (the "Software"), to deal
+// in the Software without restriction, including without limitation the rights
+// to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+// copies of the Software, and to permit persons to whom the Software is
+// furnished to do so, subject to the following conditions:
+//
+// The above copyright notice and this permission notice shall be included in
+// all copies or substantial portions of the Software.
+//
+// THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+// IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+// FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+// AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+// LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+// OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+// SOFTWARE.
+//
+// SPDX-License-Identifier: MIT
+
+// Package configgen renders a CircleCI pipeline YAML from a reposcan.Result
+// and writes it to disk. It backs `circleci config generate` and is reusable
+// by other commands (e.g. an onboarding wrapper) that already have a scan
+// result in hand and want to write the config without re-scanning.
+package configgen
+
+import (
+	"bytes"
+	"context"
+	"fmt"
+	"os"
+	"path/filepath"
+
+	"gopkg.in/yaml.v3"
+
+	clierrors "github.com/CircleCI-Public/circleci-cli/internal/errors"
+	"github.com/CircleCI-Public/circleci-cli/internal/iostream"
+	"github.com/CircleCI-Public/circleci-cli/internal/reposcan"
+)
+
+const generatedHeader = "# Generated by circleci config generate\n"
+
+// rename is the package-level seam for finalizing the atomic write. Tests
+// swap it via SetRenameForTest in export_test.go.
+var rename = os.Rename
+
+// Generate writes a CircleCI pipeline config to <dir>/.circleci/config.yml
+// from a pre-computed scan result.
+//
+// If a config file already exists at that path, Generate prints
+// "Using existing config" to stderr and returns nil — matching the
+// WEBXP-990 idempotency contract. Otherwise it prints the scan summary
+// via reposcan.Render, marshals the pipeline YAML, creates the .circleci/
+// directory if missing, writes the file atomically (temp file + rename),
+// and prints a "Generated <path>" line to stderr.
+//
+// Errors are returned as structured CLIErrors so the CLI's top-level
+// handler can render them consistently.
+func Generate(ctx context.Context, dir string, result *reposcan.Result) error {
+	configPath := filepath.Join(dir, ".circleci", "config.yml")
+	if _, err := os.Stat(configPath); err == nil {
+		iostream.ErrPrintf(ctx, "%s Using existing config at %s\n",
+			iostream.SymbolOK(ctx), configPath)
+		return nil
+	}
+
+	reposcan.Render(ctx, result)
+
+	body, err := renderConfig(result)
+	if err != nil {
+		return clierrors.New(
+			"configgen.render_failed",
+			"Could not render config",
+			fmt.Sprintf("YAML marshaling failed: %s.", err),
+		).WithExitCode(clierrors.ExitGeneralError)
+	}
+
+	if err := writeConfigAtomic(configPath, body); err != nil {
+		return clierrors.New(
+			"configgen.write_failed",
+			"Could not write config",
+			fmt.Sprintf("Failed to write %s: %s.", configPath, err),
+		).WithExitCode(clierrors.ExitGeneralError)
+	}
+
+	iostream.ErrPrintf(ctx, "%s Generated %s\n", iostream.SymbolOK(ctx), configPath)
+	return nil
+}
+
+// renderConfig builds the pipeline YAML body for a scan result. When the scan
+// produced no usable detection (IsEmpty), a generic cimg/base:stable template
+// is emitted with a placeholder build step. Output is deterministic: struct
+// field order controls top-level layout and yaml.v3 sorts map keys
+// alphabetically.
+func renderConfig(r *reposcan.Result) ([]byte, error) {
+	var image string
+	steps := []step{{Checkout: true}}
+
+	if r.IsEmpty() {
+		image = "cimg/base:stable"
+		steps = append(steps, step{Run: &runStep{
+			Name:    "build",
+			Command: `echo "Add your build steps here"`,
+		}})
+	} else {
+		image = r.Image
+		if r.ImageVersion != "" {
+			image = r.Image + ":" + r.ImageVersion
+		}
+		for _, s := range r.Setup {
+			steps = append(steps, step{Run: &runStep{Name: s.Name, Command: s.Command}})
+		}
+	}
+
+	p := pipeline{
+		Version: "2.1",
+		Jobs: map[string]pipelineJob{
+			"build": {
+				Docker: []dockerImage{{Image: image}},
+				Steps:  steps,
+			},
+		},
+		Workflows: map[string]workflow{
+			"build": {Jobs: []string{"build"}},
+		},
+	}
+
+	var buf bytes.Buffer
+	buf.WriteString(generatedHeader)
+
+	enc := yaml.NewEncoder(&buf)
+	enc.SetIndent(2)
+	if err := enc.Encode(p); err != nil {
+		return nil, err
+	}
+	if err := enc.Close(); err != nil {
+		return nil, err
+	}
+	return buf.Bytes(), nil
+}
+
+// writeConfigAtomic writes body to path via a temp file in the same directory
+// followed by rename. The .circleci directory is created if missing. If any
+// step after the temp file is created fails, the temp file is removed so the
+// .circleci/ directory is never left littered with config.*.yml.tmp files.
+func writeConfigAtomic(path string, body []byte) error {
+	if err := os.MkdirAll(filepath.Dir(path), 0o755); err != nil { //#nosec:G301 // .circleci/ is a repo-shared directory, world-readable like the surrounding workspace
+		return err
+	}
+
+	tmp, err := os.CreateTemp(filepath.Dir(path), "config.*.yml.tmp")
+	if err != nil {
+		return err
+	}
+
+	committed := false
+	defer func() {
+		if !committed {
+			_ = os.Remove(tmp.Name())
+		}
+	}()
+
+	if _, err := tmp.Write(body); err != nil {
+		_ = tmp.Close()
+		return err
+	}
+	if err := tmp.Sync(); err != nil {
+		_ = tmp.Close()
+		return err
+	}
+	if err := tmp.Close(); err != nil {
+		return err
+	}
+	if err := os.Chmod(tmp.Name(), 0o644); err != nil { //#nosec:G302 // config.yml is intended to be committed alongside the rest of .circleci/
+		return err
+	}
+	if err := rename(tmp.Name(), path); err != nil {
+		return err
+	}
+	committed = true
+	return nil
+}
+
+type pipeline struct {
+	Version   string                 `yaml:"version"`
+	Jobs      map[string]pipelineJob `yaml:"jobs"`
+	Workflows map[string]workflow    `yaml:"workflows"`
+}
+
+type pipelineJob struct {
+	Docker []dockerImage `yaml:"docker"`
+	Steps  []step        `yaml:"steps"`
+}
+
+type dockerImage struct {
+	Image string `yaml:"image"`
+}
+
+type workflow struct {
+	Jobs []string `yaml:"jobs"`
+}
+
+// step is either a bare "checkout" string or a "run" map. The custom
+// MarshalYAML selects the right form so the rendered YAML matches the
+// canonical CircleCI shape.
+type step struct {
+	Checkout bool
+	Run      *runStep
+}
+
+type runStep struct {
+	Name    string `yaml:"name"`
+	Command string `yaml:"command"`
+}
+
+func (s step) MarshalYAML() (interface{}, error) {
+	if s.Checkout {
+		return "checkout", nil
+	}
+	return map[string]*runStep{"run": s.Run}, nil
+}
