+++ internal/configgen/testdata/detected-node.stderr.golden
@@ -1,0 +1,4 @@
+✓ Detected node project (cimg/node:20.10)
+    install: npm ci
+    test: npm test
+✓ Generated <DIR>/.circleci/config.yml
