+++ internal/configgen/testdata/detected-node.yml.golden
