+++ internal/configgen/testdata/fallback.stderr.golden
