+++ internal/configgen/testdata/skip-existing.stderr.golden
@@ -1,0 +1,1 @@
+✓ Using existing config at <DIR>/.circleci/config.yml
